refactor: section title component
