+++ client/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./component";
+export * from "./form";
+export * from "./navigation";
